chore(cli-integ): add per-test timeouts (#24504)
We couldn't use the jest timeout feature for our integration tests, and ended up running without timeouts.

This made it EXTREMELY hard to debug an issue where the tests ended up not finishing, and not producing any output to indicate why.

Re-add the ability to have timeouts on test, and set a timeout of 10 minutes on all of them.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*