fix(sfn): can't override toStateJson() from other languages (#24593)
If any part of a state's JSON representation is `null`, that value will be replaced by `undefined` when jsii sends data to the other language, resulting in a change of semantics.

Multi-language APIs cannot differentiate between `null` and `undefined` as non-JS languages typically fail to distinguish between them... In order to address that, a `JsonNull` value was added which serializes to `null` (via Javascript's standard `toJSON` method), which must be used in such cases where `null` may need to cross the language boundary.

The `JsonPath.DISCARD` value is now a string-token representation of the `JsonNull` instance.

Fixes #14639
Fixes https://github.com/aws/jsii/issues/3999

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*