chore(cli): try to prevent `undefined/undefined` in user agent
It should not be possible, but I'm seeing `undefined/undefined` as the
user agent in CloudTrail logs, and it looks like it is the CLI causing
this.

Add a test (which succeeds) and some fallback behavior to help
and debug this some more if we see it again in the future.