feat(elasticloadbalancing): classic load balancer supports ec2 instances  (#24353)
Introducing the `InstanceTarget` class: an EC2 instance that serves as the target for load balancing. This class allows to register an instance to a load balancer. `InstanceTarget ` takes an instance to register as the target for the  load balancer.

For example,
```ts
const target = new elb.InstanceTarget(instance);
elb.addTarget(target);
```
creates an InstanceTarget object with the specified instance and then adds it to the load balancer.

> [CONTRIBUTING GUIDE]: https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md
> [DESIGN GUIDELINES]: https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md

Closes #23500.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*