feat: add machine readable report of recommended feature flags
`cdk init` creates a new project with feature flags that enable the
bugfixes we can't turn on by default for fear of disturbing existing
projects.

Alternative project generators that do not use `cdk init` have no way of
getting this feature flag list, so drop a machine-readable copy
of the flags into the source repository and the NPM package.

The new file will be accessible under:

```
./node_modules/@aws-cdk/cx-api/recommended-feature-flags.json
./node_modules/aws-cdk-lib/cx-api/recommended-feature-flags.json
```