feat(ec2): SSM sessions
It's not too hard to enable SSM Session Manager to Instances and
AutoScalingGroups (it's a matter of picking the right AMI and
adding the right managed policy to the instance role).

This PR adds a single boolean to turn on the policy directly and
advertises the feature in the README for people who might otherwise not
know this feature exists.