feat(ec2): CFN-init support for systemd
CFN-init can be used to install software onto EC2 Instances that are
created using CloudFormation.

CFN init supports SystemD, but this was not yet available in CDK. This
PR adds support for SystemD. It also adds a helper function to create
a simple SystemD config files for your own services.