feat(ec2): SSM sessions (#24673)
It's not too hard to enable SSM Session Manager to Instances and AutoScalingGroups (it's a matter of picking the right AMI and adding the right managed policy to the instance role).

This PR adds a single boolean to turn on the policy directly and advertises the feature in the README for people who might otherwise not know this feature exists.

Also consistentize the use and explanation of `MachineImage.latestAmazonLinux` a bit.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*