fix(toolkit): RWLock.acquireRead is not re-entrant
If multiple threads of the same process attempt to acquire the same
reader lock, the a race condition occurs, and the first thread to
release the reader lock will release ALL the locks.

Introduce a counter so that each acquire attempt uses a different file
name, ensuring that the read lock is reentrant.