fix(sfn): stop replacing JsonPath.DISCARD with `null`
Follow-up to #24593. The `renderJsonPath` function is subsituting a
literal `null` for `JsonPath.DISCARD`, which results in the key being
dropped if the value is sent across a language boundary, which
effectively changes semantics.

The `JsonPath.DISCARD` value is a `Token` that ultimately resolves to
`null`, and it must be preserved as such so that it is safe to exchange
across languages.

Thanks to @beck3905 for reporting & diagnosing this.