chore: detect property type renames to ensure no accidental backwards breakage
Upstream service teams may rename property types in the resource
specification.

This is strictly speaking not breaking from the point of view of the
spec, because the names of the property types don't appear anywhere in
the code a user would normally type (i.e., in a CloudFormation
template).

However, CDK generates classes for these types, and so the name *does*
matter and changing it is breaking.

To detect these instances, we check that during an upgrade, all old
property type names are still present. If not, the reason is probably
that they renamed a type.

Note that this is not a 100% guaranteed to catch all scenarios (I'm sure
you can think of changes that would be breaking and still pass this
check), but it's at least very likely to catch honest mistakes in
commonly expected scenarios.