fix(integ-runner): update workflow doesn't support resource replacement
For the update workflow we deploy the stack update with the
`--no-rollback` flag so that the assertion stack will deploy all the way
through instead of failing on the first assertion failure. This causes
issues if you are deploying an update that includes resource
replacement since resource replacement is not allowed when using
`--no-rollback`

To fix this, this PR splits the stack update deployment into two
separate deployments. The first deploys the test case stack and the
second deploys the assertion stack with `--no-rollback`.