feat(ecr): add option to auto delete images upon ECR repository removal (#24572)
This request fixes the ECR Repository resource to allow setting a flag on the resource to auto delete the images in the repository. This is similar to the way S3 handles the autoDeleteObjects attribute. This code base starts from a stalled PR [#15932](https://github.com/aws/aws-cdk/pull/15932). This also takes into account the functionality added into S3 to create tag to not delete images if the flag is flipped from true to false. 

Closes [#12618](https://github.com/aws/aws-cdk/issues/12618)
References closed and not merged PR  [#15932](https://github.com/aws/aws-cdk/pull/15932)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*