fix(cli): pathMetadata and assetMetadata defaults cannot be configured in cdk.json (#24533)
Previously, pathMetadata, assetMetadata, debug and staging could not be set in `cdk.json` as command-line parameters defaults would always overwrite this. In particular, this `cdk.json` would be ignored:

```
{
  "app": "npx ts-node --prefer-ts-exts bin/example",
  ...
  "pathMetadata": false,
  "assetMetadata": false
}
```

The issue is this piece of code:
```
 this.settings = this.defaultConfig
            .merge(userConfig)
            .merge(this.projectConfig)
            .merge(this.commandLineArguments)
            .makeReadOnly();
```

`commandLineArguments` returns true for the above mentioned options. This commit is a copy of https://github.com/aws/aws-cdk/pull/21891 which fell through the cracks, so I'm picking up as this can potentially cause outages to resources that have drift under the hood, and adding metadata looks like an `Update` operation to CloudFormation.

There are already unit tests around the cli which show that the same behavior works, located here:
https://github.com/aws/aws-cdk/blob/main/packages/%40aws-cdk-testing/cli-integ/tests/cli-integ-tests/cli.integtest.ts
 
shoutout to @berenddeboer who did the work, I'm just reviving it.

Closes #3573

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*