fix(aws-batch): deny zero maxvCpus in compute environment
In the batch compute environment, the maxVcpus property is
optional and defaults to 256. If users set it to zero, they
expect the compute environment to not run.

Because of Typescript's rules for truthy/falsey values, when
passing a maxVcpus value of 0, we set it to 256 in

    maxVcpu || 256

The fix is to use the null coalescing operator as in

    maxVcpu ?? 256

more importantly throw if maxvCpus is less than 1.