fix(ec2): VPC Flow Log record fields are not available (#24812)
There are a lot of VPC Flow Log record fields that are not currently
available as named attributes (from various versions). [Recently][1],
CloudWatch Logs got support for all existing versions of log record
fields. This makes sure [all][2] those available.

This also corrects an issue with `interface-id` which made it impossible
to create logs for CloudWatch Logs since it was missing the closing
brace. This is done by making use of the `field` static method for nearly
all existing fields that are just a single field. Hopefully this prevents
errors in the future.

[1]: https://aws.amazon.com/about-aws/whats-new/2023/03/amazon-cloudwatch-logs-amazon-vpc-flow-logs-metadata/
[2]: https://docs.aws.amazon.com/vpc/latest/userguide/flow-logs.html#flow-logs-fields


Closes #24807.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*