chore(ec2): KMS supports FIPS 140-2 enabled Endpoint and Other (#24679)
Since the Version Update for RDS was merged as a "feat", we will issue this Pull Request as a "feat" as well, not as a "chore". However, there is no particular preference, so it can also be issued as a "chore".

Endpoints corresponding to the following news:
https://aws.amazon.com/about-aws/whats-new/2023/03/aws-key-management-service-fips-140-2-interface-vpc-endpoints/?nc1=h_ls

Endpoints that are obviously not supported according to the following link:
https://docs.aws.amazon.com/vpc/latest/privatelink/aws-services-privatelink-support.html

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*