docs(secretsmanager): improve the doc for fromSecretNameV2() (#24877)
`fromSecretNameV2()` could lead to `AccessDeniedException` when it comes with secrets with names ending with a hyphen and 6 characters which should be avoided according to the [document](https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot.html#ARN_secretnamehyphen).

We are having relevant discussion and impacted customers in https://github.com/aws/aws-cdk/issues/18555#issuecomment-1490576142 as well. I believe it is necessary to improve the doc for this method.

Closes https://github.com/aws/aws-cdk/issues/18555

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*