fix(ec2): looking up a shared VPC has incorrect account ID in ARN (#24486)
This ensures that when a VPC is looked up using the context provider the
resulting VPC has the correct ARN. Previously, the OwnerId field in
the output was ignored; this is now passed as part of the environment of
the resource (along with its region). Additionally, owner-id as added
as a valid field to query on (as defined by the DescribeVpcs
documentation).

Further, previously the ARN did not include the region or account; this
seems to be an error as Arn.format only considers the given component
and the stack region/account, so those parameters are now passed as
well.

If OwnerId is undefined, this will fallback to the AWS::AccountId
pseudoparameter (the parent stack's Account ID); it seems like OwnerId
should always be defined though (even though the documentation
technically says it is not required).

Finally, an interface private to the tests had a typo in the name and
that is corrected (to MockVpcContextResponse).

This follows up on #23982 which was closed by automation after incorrectly
detecting it as abandoned.

Closes #23865.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*