feat(pipelines): allow artifactBucket to be overridden (#24945)
`pipelines.CodePipeline` leaves it to `aws_codepipeline.Pipeline` to create a bucket to store the artifacts. Users that use the CFN Guard plugin in a application with a pipeline will find it reporting violations on that bucket, such as the absence of logging configuration.

Allow users to override the bucket, so that they can configure it however they want or need.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*