fix(core): Add stage prefix to stack name shortening process (#24443)
There is an issue in the stack name generation process where the prefix generated from assembly's stage name is not taken into account when shortening a stack name to meet the requirement of being equal or less than 128 characters longs.

This can lead to generating an invalid stack name greater than 128 characters: since the stack name is shortened to 128 characters, when the prefix is added, the limit is exceeded.

Current solution:
- Add the prefix into the first ID if any prefix is present.

Alternative solution:
- Add the prefix at the start of the ID list. Doing this would require to add more conditionals and consider more corner cases to keep the names consistent with the old algorithm.
 

Fixes #23628

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*