chore: switch cli-integ-testing package to stable (#25131)
The scripts that install the testing package expect it to have a version number that is stable, but because its stability is `experimental` it gets the alpha version treatment.

Error during pipeline build:

```
npm ERR! notarget No matching version found for @aws-cdk-testing/cli-integ@2.74.0-rc.0.
npm ERR! notarget In most cases you or one of your dependencies are requesting
```

But the actual version in the repository is:

```
2.74.0-alpha.999
```

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*