fix(rds): Correct ARN in IAM policy for IAM database access
The IAM policy for IAM database access takes the form of:

```
arn:aws:rds-db:region:account-id:dbuser:DbiResourceId/db-user-name
```

Update the ARN used in `grantConnect` to this format.

Additionally, update the signature of `grantConnect` to take an optional `dbUser`,
which is defaulted to the master username of the database.

This signature change also matches `grantConnect` in `DatabaseProxy`.

See:
 - https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.IAMDBAuth.IAMPolicy.html
 - https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/105