fix(servicecatalogappregistry): Remove deprecated resource in Application Construct (#25095)
BREAKING CHANGE: this change will deprecated **associateStack** and **associateAttributeGroup** in Application Construct. 
The user who are using these two method need to update to use alternative method.
For associateStack, the alternative method is **associateApplicationWithStack**
For associateAttributeGroup, the alternative method is **AttributeGroup.associateWith**

The user who are using these two method need to update to use alternative method. For associateStack, the alternative method is **associateApplicationWithStack** For associateAttributeGroup, the alternative method is **AttributeGroup.associateWith**

Purpose of this PR:

we need to remove deprecated resource before we moving into stable version The method that we remove is: associateStack and associateAttributeGroup

CHANGES:

1. in lib/application.ts, we remove these two methods and update their corresponding interface
2. in test/ application.test.ts & test/integ.application.ts, we update application.test.ts and integ.application.ts to remove these two methods' related test


*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*