fix(aws-cdk-lib): compiled .js files are no longer being minified (#25160)
Prior to the aws-cdk repo remodel, there was a script executed at the end of build to minify the `.js` files. After the remodel, that script was excluded. In this change, I am adding it back in as a pre-package step. There is no need to perform this as part of the build. The `minify-sources` script takes about 2 seconds to run. 

The size of aws-cdk-lib v 2.74 is 222 MB. This change will bring the overall size down to 205 MB. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*