fix(bootstrap): add `previous-parameters` option to bootstrap command
A recent change was made to explicitly set the `usePreviousParameters`
when deploying the bootstrap template. This is because if a template is
deployed once by a user with a set of parameters and then later by
another user we shouldn't drop those previously set parameters.

This PR adds the option for the user to control this behavior. If the
user needs to turn this off they can provide the
`--no-previous-parameters` option.

fixes #23780