feat(ecs-patterns): Tagging support for scheduled tasks (#25222)
This was enabled by PR #23838 which adds support for the `propagateTags` and `tags` attributes to the ecs-patterns module. With the changes in this PR, tags can now be applied to tasks that are launched as scheduled tasks with the proper tags passed through to the event target.

This PR contributes to (but does not completely close) #25106 .

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*