feat(pipelines): added logging as option for codeBuildDefaults prop on CodePipeline construct (#25266)
## Motivation:
Logging is an optional prop as part of the Project construct in the aws-codebuild module. Users can link a codebuild action when as part of their CodePipeline with codebuild options being set via the codeBuildDefaults prop on the CodePipeline construct. Currently, there isn't any ability to enable logging for the codebuild action. This feature allows a user to enable logging as part of a linked codebuild action via S3, CloudWatch, or both.

## Use Case (from issue #22045):
At my company we just enabled AWS Config and SecurityHub has generated a rule 'securityhub-codebuild-project-logging-enabled-abcxyz'. This rule states that all CodeBuild projects must have at least one logging option enabled.

However the Build and Synth steps from the pipeline are always failing this check since the pipelines module doesn't allow specifying logging options on the nested CodeBuild projects.

## Testing:
Unit testing was implemented to verify that the CloudFormation template would be synthesized correctly when logging selected as an option on the codeBuildDefaults prop for S3, CloudWatch, or both.

Integration testing was implemented to verify that all resources would be successfully provisioned when a new CodePipeline is created with logging enabled for the codebuild action.

Closes #22045

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*