feat(ec2): restrict access to default security group (under feature flag)
This PR implements functionality which will remove the default
ingress/egress rules from the VPC default security group. When a VPC is
created, the default security group is created as well with default
ingress/egress rules which allow _all_ traffic. It is not possible to
delete the default security group, but you should never use it. As a
result there are a log of security standards that recommend removing the
default rules so that the security group denies all traffic by default.
See [this
rule](https://docs.aws.amazon.com/securityhub/latest/userguide/ec2-controls.html#ec2-2).

Since the default security group cannot be managed through a
CloudFormation resource, this PR introduces a new Custom Resource which
will remove the ingress/egress rules.

I also think that this should be the default behavior so I have
introduced a new feature flag to make this the default for new apps. As
a result I had to update _a lot_ of integration tests. Since This
feature flag would only be introduced on new VPCs it didn't make sense
to run the update workflow on all these integration tests so I updated
them to disable this new feature.

I added one new integration test to test this functionality.

fixes #19394