fix: log buckets don't have acls enabled
Adds `objectOwnership` to s3 server access log delivery buckets that are
created in cdk applications with the `@aws-cdk/aws-s3:serverAccessLogsUseBucketPolicy` feature flag disabled. This will allow users to keep creating new buckets within those apps for storing logs. This is only set if the user has not configured `objectOwnership` manually on the log bucket.

`ObjectWriter` was essentially the default behavior before the change to
disable ACLs by default for new buckets so though this will update
existing buckets it should not cause any breakage or replacement.