fix(cli): diff doesn't display paths for removed resources (#25294)
Even if the old template has paths in it, they are not displayed because all paths are loaded from the assembly metadata.

Include the logicalID paths from the old template, if present, but give preference to the new template.

This will show paths for removed resources, if there are any. Useful to debug replacements.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*