fix(pipelines): CodeBuild Action role can be assumed by too many identities (#25318)
Backport of https://github.com/aws/aws-cdk/pull/25316.

CDK Pipelines creates a single Role which has permissions to start all CodeBuild jobs. The AssumeRolePolicy for this Role contained a mistake, which allowed all roles in the same account with appropriate sts:AssumeRole permissions to assume the Role.

Fix this by limiting the AssumeRolePolicy to the actual pipeline's execution role, which we have so we can reference directly.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*