chore(aws-cdk-lib): Run pre and post package steps for aws-cdk-lib from pack.sh (#25239)
Until this PR, the pre and post package steps defined in `aws-cdk-lib`'s `package.json` were not actually executed during the pack step of the pipeline. The ultimate purpose of this change is to make sure that the `.js` source files are minified in the release pipeline, resulting in a smaller overall package size. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*