fix(integ-tests): allow multiple AwsApiCalls with the same action and different parameters (#25241)
This PR allows multiple AwsApiCalls with the same action and different parameters.
Closes #25014

To be able to test the new behavior, a new integration test has been added in aws-redshift-alpha package.

Next possible step (to be discussed): add an optional `callNumber` in `awsApiCall` method to be able to use this method twice with the same action and same parameters.