chore(pkglint): ignore alpha version of `cdk-cli-wrapper` (#25382)
When we check that all dependency version ranges are correct in `pkglint`, we don't account for alpha modules.

That is usually fine because no one should be depending on alpha modules, but the `integ-runner` depends on the package `cdk-cli-wrapper` which is alpha.

Add an exception (that's simpler than redoing everything). This fixes the build.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*