fix(core): crossRegionReferences don't work across multiple regions
The first attempt to fix this in #25190 didn't work because it didn't
account for the fact that when exporting to multiple regions, we create
multiple `ExportWriter`s that all use the same provider (and provider
role).

This PR fixes that by adding the policy cross region ARNs directly to
the custom resource provider (1 per stack) rather than the
`ExportWriter` (multiple per stack). I also updated the test case to
better account for this scenario.

fixes #25377