fix(api-gateway): add validation to `variables` property on `Stage` resource (#25267)
Closes #3635 

From the original issue: API Gateway Stage variables must match a regex to be valid ([See the documentation](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-apigateway-stage.html#cfn-apigateway-stage-variables)). When adding variables using CDK, this regex is not enforced, so it can synthesize a CloudFormation template that cannot be deployed.

This PR fixes this issue by adding a validation that the strings match that regex along with a unit test for these changes.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*