chore(prlint): use `pr/needs-review` label
It is currently very difficult to use the issue search functionality to
figure out which PRs are ready for review from a core team member.
Usually it is ~ 1/3 of open PRs that are in a reviewable state.

Some of the information we need is not available as a search filter in
the issue search page (or it is difficult to create a correct filter).
A PR is ready for review when:

  1. Not a draft
  2. Does not have any merge conflicts
  3. PR linter is not failing OR the user has requested an exemption
  4. A maintainer has not requested changes
  ...others may be added in the future

This PR adds functionality to add a label `pr/needs-review` to any PR
that meets our criteria. This will allow us to easily use this to filter
PRs.