chore(aws-cdk-lib): ignore cfn-types-2-classes.json
This is a generated file that does not need to be present in the repo.

Also changed the ordering of services.
Previously this was ordered by UTF-16 code unit values.
Now it's human friendly, in case anyone ever needs to look at it.
This also helps with the transition to the new l1 generation.