chore(aws-cdk-lib): ignore cfn-types-2-classes.json (#25436)
This is a generated file that does not need to be present in the repo.

Also changed the ordering of services.
Previously this was ordered by UTF-16 code unit values. Now it's human friendly, in case anyone ever needs to look at it. This also helps with the transition to the new l1 generation.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*