chore(prlint): remove needs-review label on approved prs (#25506)
I figure that approved PRs should also lose the `needs-review` label, since they will still exist in the open PR list for a while and pollute the pool of actual prs needing review. I could've bundled this in to `maintainerRequestsChanges` but I figured it makes more sense to be explicit about the different rules we have for the `needs-review` label.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*